requirements.txt
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
sentence-transformers
Flask
requests
gunicorn